Introduce IEntityDao and EntityHibernateDao
Summary:
Entity就是一个能被create, udpate, delete, findById的对象。

IEntityDao定义了能对Entity进行的操作，EntityHibernateDao定义了基于hibernate的实现

这2个类能减少大量的重复代码

Test Plan: 和canming联合疯狂测试

Reviewers: Seanjiang, gaoxh04

Reviewed By: gaoxh04

Differential Revision: http://phabricator.dasinong666.com/D64

+++ src/main/java/com/dasinong/ploughHelper/inputParser/UserParser.java
@@ -14,9 +14,9 @@ public class UserParser {
         String address = request.getParameter("address");
         String deviceId = request.getParameter("deviceId");
         String channel = request.getParameter("channel");
-        int institutionId = 0;
+        Long institutionId = 0L;
         try{
-        	institutionId = Integer.parseInt(request.getParameter("institutionId"));
+        	institutionId = Long.parseLong(request.getParameter("institutionId"));
         }catch(Exception e){
         	System.out.println("Inproper institutionId. Use 0 as default");
         }
