+++ src/main/java/com/dasinong/ploughHelper/outputWrapper/UserWrapper.java
@@ -23,7 +23,7 @@ public class UserWrapper implements Serializable {
 	private String channel;
 	private String refcode;
 	private Long refuid;
-	private int institutionId;
+	private Long institutionId;
 	
 	private List<Long> fields = new ArrayList<Long>();
 	private List<Integer> monitorLocationId = new ArrayList<Integer>();
@@ -140,10 +140,10 @@ public class UserWrapper implements Serializable {
 	public void setRefuid(Long refuid) {
 		this.refuid = refuid;
 	}
-	public int getInstitutionId() {
+	public Long getInstitutionId() {
 		return institutionId;
 	}
-	public void setInstitutionId(int institutionId) {
+	public void setInstitutionId(Long institutionId) {
 		this.institutionId = institutionId;
 	}
